Fixes for archive imports (#55)
* Find data files in subdirectories in archives

See #54 for discussion. This also ignores files in the new "assets"
directory, which appear to be some stuff for a browser interface
Twitter's created.

* Fix list-member importer

It appears in list data that some rows contain a `urls` key with a list
of URLs, while others contain a `url` key with just a single one. This
change supports either way.

* Fix tweet import

This was working, sorta, but wasn't properly unpacking the tweet
data into columns. This commit fixes that in what I think should
be a backwards-compatible way.